CMP-519 Each root resource should create and store a unique id (UID).

+++ src/main/test/org/compass/core/test/nounmarshall/simple/A.java
@@ -21,7 +21,7 @@ package org.compass.core.test.nounmarshall.simple;
  */
 public class A {
 
-    Long id;
+    int id;
 
     String value;
